pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "openai>=1.101.0",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "sqlglot>=27.8.0",
    "sqlparse>=0.5.3",
    "streamlit>=1.48.1",
    "uvicorn>=0.35.0",
]
